fix: fix error conflict

+++ src/CloseExpAISolution.Domain/Entities/OrderItem.cs
@@ -5,11 +5,10 @@ public class OrderItem
     public Guid OrderItemId { get; set; }
     public Guid OrderId { get; set; }
     public Guid LotId { get; set; }
-    public int Quantity { get; set; }
+    public short Quantity { get; set; }
     public decimal UnitPrice { get; set; }
     public decimal TotalPrice { get; set; }
 
     public Order? Order { get; set; }
     public StockLot? StockLot { get; set; }
 }
-
